+++ pcsx2/IPU/IPU.cpp
@@ -17,8 +17,8 @@
 #include "Common.h"
 
 #include "IPU.h"
+#include "IPU_MultiISA.h"
 #include "IPUdma.h"
-#include "yuv2rgb.h"
 #include "mpeg2lib/Mpeg.h"
 
 #include <limits.h>
@@ -31,7 +31,7 @@ alignas(16) tIPU_cmd ipu_cmd;
 alignas(16) tIPU_BP g_BP;
 alignas(16) decoder_t decoder;
 
-void IPUWorker();
+static void (*IPUWorker)();
 
 // Color conversion stuff, the memory layout is a total hack
 // convert_data_buffer is a pointer to the internal rgb struct (the first param in convert_init_t)
@@ -40,11 +40,11 @@ void IPUWorker();
 //u8 PCT[] = {'r', 'I', 'P', 'B', 'D', '-', '-', '-'};		// unused?
 
 // Quantization matrix
-static rgb16_t vqclut[16];			//clut conversion table
-static u16 s_thresh[2];				//thresholds for color conversions
+rgb16_t g_ipu_vqclut[16]; //clut conversion table
+u16 g_ipu_thresh[2]; //thresholds for color conversions
 int coded_block_pattern = 0;
 
-alignas(16) static u8 indx4[16*16/2];
+alignas(16) u8 g_ipu_indx4[16*16/2];
 
 uint eecount_on_last_vdec = 0;
 bool FMVstarted = false;
@@ -67,6 +67,7 @@ __fi void IPUProcessInterrupt()
 
 void ipuReset()
 {
+	IPUWorker = MULTI_ISA_SELECT(IPUWorker);
 	memzero(ipuRegs);
 	memzero(g_BP);
 	memzero(decoder);
@@ -84,8 +85,8 @@ void ReportIPU()
 	Console.WriteLn(ipu_fifo.in.desc());
 	Console.WriteLn(ipu_fifo.out.desc());
 	Console.WriteLn(g_BP.desc());
-	Console.WriteLn("vqclut = 0x%x.", vqclut);
-	Console.WriteLn("s_thresh = 0x%x.", s_thresh);
+	Console.WriteLn("vqclut = 0x%x.", g_ipu_vqclut);
+	Console.WriteLn("thresh = 0x%x.", g_ipu_thresh);
 	Console.WriteLn("coded_block_pattern = 0x%x.", coded_block_pattern);
 	Console.WriteLn("g_decoder = 0x%x.", &decoder);
 	Console.WriteLn("mpeg2_scan = 0x%x.", &mpeg2_scan);
@@ -101,8 +102,8 @@ void SaveStateBase::ipuFreeze()
 	Freeze(ipu_fifo);
 
 	Freeze(g_BP);
-	Freeze(vqclut);
-	Freeze(s_thresh);
+	Freeze(g_ipu_vqclut);
+	Freeze(g_ipu_thresh);
 	Freeze(coded_block_pattern);
 	Freeze(decoder);
 	Freeze(ipu_cmd);
@@ -408,304 +409,12 @@ static __ri void ipuBDEC(tIPU_CMD_BDEC bdec)
 	memzero_sse_a(decoder.mb16);
 }
 
-static __fi bool ipuVDEC(u32 val)
-{
-	static int count = 0;
-	if (count++ > 5) {
-		if (!FMVstarted) {
-			EnableFMV = true;
-			FMVstarted = true;
-		}
-		count = 0;
-	}
-	eecount_on_last_vdec = cpuRegs.cycle;
-
-	switch (ipu_cmd.pos[0])
-	{
-		case 0:
-			if (!bitstream_init()) return false;
-
-			switch ((val >> 26) & 3)
-			{
-				case 0://Macroblock Address Increment
-					decoder.mpeg1 = ipuRegs.ctrl.MP1;
-					ipuRegs.cmd.DATA = get_macroblock_address_increment();
-					break;
-
-				case 1://Macroblock Type
-					decoder.frame_pred_frame_dct = 1;
-					decoder.coding_type = ipuRegs.ctrl.PCT > 0 ? ipuRegs.ctrl.PCT : 1; // Kaiketsu Zorro Mezase doesn't set a Picture type, seems happy with I
-					ipuRegs.cmd.DATA = get_macroblock_modes();
-					break;
-
-				case 2://Motion Code
-					ipuRegs.cmd.DATA = get_motion_delta(0);
-					break;
-
-				case 3://DMVector
-					ipuRegs.cmd.DATA = get_dmv();
-					break;
-
-				jNO_DEFAULT
-			}
-
-			// HACK ATTACK!  This code OR's the MPEG decoder's bitstream position into the upper
-			// 16 bits of DATA; which really doesn't make sense since (a) we already rewound the bits
-			// back into the IPU internal buffer above, and (b) the IPU doesn't have an MPEG internal
-			// 32-bit decoder buffer of its own anyway.  Furthermore, setting the upper 16 bits to
-			// any value other than zero appears to work fine.  When set to zero, however, FMVs run
-			// very choppy (basically only decoding/updating every 30th frame or so). So yeah,
-			// someone with knowledge on the subject please feel free to explain this one. :) --air
-
-			// The upper bits are the "length" of the decoded command, where the lower is the address.
-			// This is due to differences with IPU and the MPEG standard. See get_macroblock_address_increment().
-
-			ipuRegs.ctrl.ECD = (ipuRegs.cmd.DATA == 0);
-			[[fallthrough]];
-
-		case 1:
-			if (!getBits32((u8*)&ipuRegs.top, 0))
-			{
-				ipu_cmd.pos[0] = 1;
-				return false;
-			}
-
-			ipuRegs.top = BigEndian(ipuRegs.top);
-
-			IPU_LOG("VDEC command data 0x%x(0x%x). Skip 0x%X bits/Table=%d (%s), pct %d",
-			        ipuRegs.cmd.DATA, ipuRegs.cmd.DATA >> 16, val & 0x3f, (val >> 26) & 3, (val >> 26) & 1 ?
-			        ((val >> 26) & 2 ? "DMV" : "MBT") : (((val >> 26) & 2 ? "MC" : "MBAI")), ipuRegs.ctrl.PCT);
-
-			return true;
-
-		jNO_DEFAULT
-	}
-
-	return false;
-}
-
-static __ri bool ipuFDEC(u32 val)
-{
-	if (!getBits32((u8*)&ipuRegs.cmd.DATA, 0)) return false;
-
-	ipuRegs.cmd.DATA = BigEndian(ipuRegs.cmd.DATA);
-	ipuRegs.top = ipuRegs.cmd.DATA;
-
-	IPU_LOG("FDEC read: 0x%08x", ipuRegs.top);
-
-	return true;
-}
-
-static bool ipuSETIQ(u32 val)
-{
-	if ((val >> 27) & 1)
-	{
-		u8 (&niq)[64] = decoder.niq;
-
-		for(;ipu_cmd.pos[0] < 8; ipu_cmd.pos[0]++)
-		{
-			if (!getBits64((u8*)niq + 8 * ipu_cmd.pos[0], 1)) return false;
-		}
-
-		IPU_LOG("Read non-intra quantization matrix from FIFO.");
-		for (uint i = 0; i < 8; i++)
-		{
-			IPU_LOG("%02X %02X %02X %02X %02X %02X %02X %02X",
-			        niq[i * 8 + 0], niq[i * 8 + 1], niq[i * 8 + 2], niq[i * 8 + 3],
-			        niq[i * 8 + 4], niq[i * 8 + 5], niq[i * 8 + 6], niq[i * 8 + 7]);
-		}
-	}
-	else
-	{
-		u8 (&iq)[64] = decoder.iq;
-
-		for(;ipu_cmd.pos[0] < 8; ipu_cmd.pos[0]++)
-		{
-			if (!getBits64((u8*)iq + 8 * ipu_cmd.pos[0], 1)) return false;
-		}
-
-		IPU_LOG("Read intra quantization matrix from FIFO.");
-		for (uint i = 0; i < 8; i++)
-		{
-			IPU_LOG("%02X %02X %02X %02X %02X %02X %02X %02X",
-			        iq[i * 8 + 0], iq[i * 8 + 1], iq[i * 8 + 2], iq[i *8 + 3],
-			        iq[i * 8 + 4], iq[i * 8 + 5], iq[i * 8 + 6], iq[i *8 + 7]);
-		}
-	}
-
-	return true;
-}
-
-static bool ipuSETVQ(u32 val)
-{
-	for(;ipu_cmd.pos[0] < 4; ipu_cmd.pos[0]++)
-	{
-		if (!getBits64(((u8*)vqclut) + 8 * ipu_cmd.pos[0], 1)) return false;
-	}
-
-	IPU_LOG("SETVQ command.   Read VQCLUT table from FIFO.\n"
-	    "%02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d\n"
-	    "%02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d\n"
-	    "%02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d\n"
-	    "%02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d %02d:%02d:%02d",
-	    vqclut[0].r, vqclut[0].g, vqclut[0].b,
-	    vqclut[1].r, vqclut[1].g, vqclut[1].b,
-	    vqclut[2].r, vqclut[2].g, vqclut[2].b,
-	    vqclut[3].r, vqclut[3].g, vqclut[3].b,
-	    vqclut[4].r, vqclut[4].g, vqclut[4].b,
-	    vqclut[5].r, vqclut[5].g, vqclut[5].b,
-	    vqclut[6].r, vqclut[6].g, vqclut[6].b,
-	    vqclut[7].r, vqclut[7].g, vqclut[7].b,
-	    vqclut[8].r, vqclut[8].g, vqclut[8].b,
-	    vqclut[9].r, vqclut[9].g, vqclut[9].b,
-	    vqclut[10].r, vqclut[10].g, vqclut[10].b,
-	    vqclut[11].r, vqclut[11].g, vqclut[11].b,
-	    vqclut[12].r, vqclut[12].g, vqclut[12].b,
-	    vqclut[13].r, vqclut[13].g, vqclut[13].b,
-	    vqclut[14].r, vqclut[14].g, vqclut[14].b,
-	    vqclut[15].r, vqclut[15].g, vqclut[15].b);
-
-	return true;
-}
-
-// IPU Transfers are split into 8Qwords so we need to send ALL the data
-static __ri bool ipuCSC(tIPU_CMD_CSC csc)
-{
-	csc.log_from_YCbCr();
-
-	for (;ipu_cmd.index < (int)csc.MBC; ipu_cmd.index++)
-	{
-		for(;ipu_cmd.pos[0] < 48; ipu_cmd.pos[0]++)
-		{
-			if (!getBits64((u8*)&decoder.mb8 + 8 * ipu_cmd.pos[0], 1)) return false;
-		}
-
-		ipu_csc(decoder.mb8, decoder.rgb32, 0);
-		if (csc.OFM) ipu_dither(decoder.rgb32, decoder.rgb16, csc.DTE);
-
-		if (csc.OFM)
-		{
-			ipu_cmd.pos[1] += ipu_fifo.out.write(((u32*) & decoder.rgb16) + 4 * ipu_cmd.pos[1], 32 - ipu_cmd.pos[1]);
-			if (ipu_cmd.pos[1] < 32) return false;
-		}
-		else
-		{
-			ipu_cmd.pos[1] += ipu_fifo.out.write(((u32*) & decoder.rgb32) + 4 * ipu_cmd.pos[1], 64 - ipu_cmd.pos[1]);
-			if (ipu_cmd.pos[1] < 64) return false;
-		}
-
-		ipu_cmd.pos[0] = 0;
-		ipu_cmd.pos[1] = 0;
-	}
-
-	return true;
-}
-
-static __ri bool ipuPACK(tIPU_CMD_CSC csc)
-{
-	csc.log_from_RGB32();
-
-	for (;ipu_cmd.index < (int)csc.MBC; ipu_cmd.index++)
-	{
-		for(;ipu_cmd.pos[0] < (int)sizeof(macroblock_rgb32) / 8; ipu_cmd.pos[0]++)
-		{
-			if (!getBits64((u8*)&decoder.rgb32 + 8 * ipu_cmd.pos[0], 1)) return false;
-		}
-
-		ipu_dither(decoder.rgb32, decoder.rgb16, csc.DTE);
-
-		if (!csc.OFM) ipu_vq(decoder.rgb16, indx4);
-
-		if (csc.OFM)
-		{
-			ipu_cmd.pos[1] += ipu_fifo.out.write(((u32*) & decoder.rgb16) + 4 * ipu_cmd.pos[1], 32 - ipu_cmd.pos[1]);
-			if (ipu_cmd.pos[1] < 32) return false;
-		}
-		else
-		{
-			ipu_cmd.pos[1] += ipu_fifo.out.write(((u32*)indx4) + 4 * ipu_cmd.pos[1], 8 - ipu_cmd.pos[1]);
-			if (ipu_cmd.pos[1] < 8) return false;
-		}
-
-		ipu_cmd.pos[0] = 0;
-		ipu_cmd.pos[1] = 0;
-	}
-
-	return true;
-}
-
 static void ipuSETTH(u32 val)
 {
-	s_thresh[0] = (val & 0x1ff);
-	s_thresh[1] = ((val >> 16) & 0x1ff);
+	g_ipu_thresh[0] = (val & 0x1ff);
+	g_ipu_thresh[1] = ((val >> 16) & 0x1ff);
 	IPU_LOG("SETTH (Set threshold value)command %x.", val&0x1ff01ff);
 }
-
-// --------------------------------------------------------------------------------------
-//  CORE Functions (referenced from MPEG library)
-// --------------------------------------------------------------------------------------
-__fi void ipu_csc(macroblock_8& mb8, macroblock_rgb32& rgb32, int sgn)
-{
-	int i;
-	u8* p = (u8*)&rgb32;
-
-	yuv2rgb();
-
-	if (s_thresh[0] > 0)
-	{
-		for (i = 0; i < 16*16; i++, p += 4)
-		{
-			if ((p[0] < s_thresh[0]) && (p[1] < s_thresh[0]) && (p[2] < s_thresh[0]))
-				*(u32*)p = 0;
-			else if ((p[0] < s_thresh[1]) && (p[1] < s_thresh[1]) && (p[2] < s_thresh[1]))
-				p[3] = 0x40;
-		}
-	}
-	else if (s_thresh[1] > 0)
-	{
-		for (i = 0; i < 16*16; i++, p += 4)
-		{
-			if ((p[0] < s_thresh[1]) && (p[1] < s_thresh[1]) && (p[2] < s_thresh[1]))
-				p[3] = 0x40;
-		}
-	}
-	if (sgn)
-	{
-		for (i = 0; i < 16*16; i++, p += 4)
-		{
-			*(u32*)p ^= 0x808080;
-		}
-	}
-}
-
-__fi void ipu_vq(macroblock_rgb16& rgb16, u8* indx4)
-{
-	const auto closest_index = [&](int i, int j) {
-		u8 index = 0;
-		int min_distance = std::numeric_limits<int>::max();
-		for (u8 k = 0; k < 16; ++k)
-		{
-			const int dr = rgb16.c[i][j].r - vqclut[k].r;
-			const int dg = rgb16.c[i][j].g - vqclut[k].g;
-			const int db = rgb16.c[i][j].b - vqclut[k].b;
-			const int distance = dr * dr + dg * dg + db * db;
-
-			// XXX: If two distances are the same which index is used?
-			if (min_distance > distance)
-			{
-				index = k;
-				min_distance = distance;
-			}
-		}
-
-		return index;
-	};
-
-	for (int i = 0; i < 16; ++i)
-		for (int j = 0; j < 8; ++j)
-			indx4[i * 8 + j] = closest_index(i, 2 * j + 1) << 4 | closest_index(i, 2 * j);
-}
-
 
 // --------------------------------------------------------------------------------------
 //  Buffer reader
@@ -902,7 +611,7 @@ __fi void IPUCMD_WRITE(u32 val)
 			break;
 
 		jNO_DEFAULT;
-			}
+	}
 
 	ipuRegs.ctrl.BUSY = 1;
 
@@ -915,72 +624,4 @@ __fi void IPUCMD_WRITE(u32 val)
 	}
 	else
 		IPUWorker();
-}
-
-__noinline void IPUWorker()
-{
-	pxAssert(ipuRegs.ctrl.BUSY);
-
-	switch (ipu_cmd.CMD)
-	{
-		// These are unreachable (BUSY will always be 0 for them)
-		//case SCE_IPU_BCLR:
-		//case SCE_IPU_SETTH:
-			//break;
-
-		case SCE_IPU_IDEC:
-			if (!mpeg2sliceIDEC()) return;
-
-			//ipuRegs.ctrl.OFC = 0;
-			ipuRegs.topbusy = 0;
-			ipuRegs.cmd.BUSY = 0;
-			break;
-
-		case SCE_IPU_BDEC:
-			if (!mpeg2_slice()) return;
-
-			ipuRegs.topbusy = 0;
-			ipuRegs.cmd.BUSY = 0;
-
-			//if (ipuRegs.ctrl.SCD || ipuRegs.ctrl.ECD) hwIntcIrq(INTC_IPU);
-			break;
-
-		case SCE_IPU_VDEC:
-			if (!ipuVDEC(ipu_cmd.current)) return;
-
-			ipuRegs.topbusy = 0;
-			ipuRegs.cmd.BUSY = 0;
-			break;
-
-		case SCE_IPU_FDEC:
-			if (!ipuFDEC(ipu_cmd.current)) return;
-
-			ipuRegs.topbusy = 0;
-			ipuRegs.cmd.BUSY = 0;
-			break;
-
-		case SCE_IPU_SETIQ:
-			if (!ipuSETIQ(ipu_cmd.current)) return;
-			break;
-
-		case SCE_IPU_SETVQ:
-			if (!ipuSETVQ(ipu_cmd.current)) return;
-			break;
-
-		case SCE_IPU_CSC:
-			if (!ipuCSC(ipu_cmd.current)) return;
-			break;
-
-		case SCE_IPU_PACK:
-			if (!ipuPACK(ipu_cmd.current)) return;
-			break;
-
-		jNO_DEFAULT
-			}
-
-	// success
-	IPU_LOG("IPU Command finished");
-	ipuRegs.ctrl.BUSY = 0;
-	//ipu_cmd.current = 0xffffffff;
-	hwIntcIrq(INTC_IPU);
 }
