+++ pcsx2/pcsx2.vcxproj.filters
@@ -692,6 +692,9 @@
     <ClCompile Include="IPU\IPU_Fifo.cpp">
       <Filter>System\Ps2\IPU</Filter>
     </ClCompile>
+    <ClCompile Include="IPU\IPU_MultiISA.cpp">
+      <Filter>System\Ps2\IPU</Filter>
+    </ClCompile>
     <ClCompile Include="IPU\yuv2rgb.cpp">
       <Filter>System\Ps2\IPU</Filter>
     </ClCompile>
@@ -2009,6 +2012,9 @@
       <Filter>System\Ps2\IPU</Filter>
     </ClInclude>
     <ClInclude Include="IPU\IPU_Fifo.h">
+      <Filter>System\Ps2\IPU</Filter>
+    </ClInclude>
+    <ClInclude Include="IPU\IPU_MultiISA.h">
       <Filter>System\Ps2\IPU</Filter>
     </ClInclude>
     <ClInclude Include="IPU\yuv2rgb.h">
